+++ Providers/FreeIPA/Private/Test-FreeIPAPrerequisite.ps1
@@ -50,7 +50,8 @@ function Test-FreeIPAPrerequisite {
             'FreeIPAPermissions.csv', 'FreeIPAPrivileges.csv', 'FreeIPARoles.csv', 'FreeIPAPasswordPolicies.csv',
             'FreeIPAServices.csv', 'FreeIPAServiceDelegation.csv', 'FreeIPAIdViews.csv', 'FreeIPAIdOverrides.csv',
             'FreeIPAOtpTokens.csv', 'FreeIPAAutomemberRules.csv', 'FreeIPAAutomount.csv', 'FreeIPASelinuxUserMaps.csv',
-            'FreeIPACertMapRules.csv', 'FreeIPACaAcls.csv', 'FreeIPACertificates.csv', 'FreeIPADnsRecords.csv'
+            'FreeIPACertMapRules.csv', 'FreeIPACaAcls.csv', 'FreeIPACertificates.csv', 'FreeIPADnsRecords.csv',
+            'FreeIPAIdentityProviders.csv'
         )
         foreach ($file in $requiredFiles) {
             $fullPath = Join-Path -Path $dataPath -ChildPath $file